chore(routes): add dedicated route registration service
Decouples router from route registration
