+++ engine/classes/Elgg/Router.php
@@ -4,6 +4,7 @@ namespace Elgg;
 
 use Elgg\Http\Request;
 use Elgg\Http\ResponseBuilder;
+use Elgg\Http\ResponseFactory;
 use Elgg\Router\Middleware\WalledGarden;
 use Elgg\Router\Route;
 use Elgg\Router\RouteCollection;
@@ -48,14 +49,14 @@ class Router {
 	protected $matcher;
 
 	/**
-	 * @var UrlGenerator
-	 */
-	protected $generator;
-
-	/**
 	 * @var HandlersService
 	 */
 	protected $handlers;
+
+	/**
+	 * @var ResponseFactory
+	 */
+	protected $response;
 
 	/**
 	 * @var Route
@@ -65,24 +66,24 @@ class Router {
 	/**
 	 * Constructor
 	 *
-	 * @param PluginHooksService $hooks     Hook service
-	 * @param RouteCollection    $routes    Route collection
-	 * @param UrlMatcher         $matcher   URL Matcher
-	 * @param UrlGenerator       $generator URL Generator
-	 * @param HandlersService    $handlers  Handlers service
+	 * @param PluginHooksService $hooks    Hook service
+	 * @param RouteCollection    $routes   Route collection
+	 * @param UrlMatcher         $matcher  URL Matcher
+	 * @param HandlersService    $handlers Handlers service
+	 * @param ResponseFactory    $response Response
 	 */
 	public function __construct(
 		PluginHooksService $hooks,
 		RouteCollection $routes,
 		UrlMatcher $matcher,
-		UrlGenerator $generator,
-		HandlersService $handlers
+		HandlersService $handlers,
+		ResponseFactory $response
 	) {
 		$this->hooks = $hooks;
 		$this->routes = $routes;
 		$this->matcher = $matcher;
-		$this->generator = $generator;
 		$this->handlers = $handlers;
+		$this->response = $response;
 	}
 
 	/**
@@ -99,6 +100,10 @@ class Router {
 	 * @access private
 	 */
 	public function route(Request $request) {
+		if ($this->timer) {
+			$this->timer->begin(['build page']);
+		}
+
 		$segments = $request->getUrlSegments();
 		if ($segments) {
 			$identifier = array_shift($segments);
@@ -114,10 +119,6 @@ class Router {
 			'handler' => $identifier, // backward compatibility
 			'segments' => $segments,
 		];
-
-		if ($this->timer) {
-			$this->timer->begin(['build page']);
-		}
 
 		try {
 			$result = $old;
@@ -171,7 +172,7 @@ class Router {
 
 					$parameters['_url'] = $url;
 					$parameters['_path'] = $path;
-					
+
 					$this->current_route->setMatchedParameters($parameters);
 
 					foreach ($parameters as $key => $value) {
@@ -232,217 +233,6 @@ class Router {
 	}
 
 	/**
-	 * Register a function that gets called when the first part of a URL is
-	 * equal to the identifier.
-	 *
-	 * @param string $identifier The page type to handle
-	 * @param string $function   Your function name
-	 *
-	 * @return bool Depending on success
-	 * @deprecated 3.0
-	 */
-	public function registerPageHandler($identifier, $function) {
-		if (!is_callable($function, true)) {
-			return false;
-		}
-
-		$this->registerRoute($identifier, [
-			'path' => "/$identifier/{segments}",
-			'handler' => $function,
-			'defaults' => [
-				'segments' => '',
-			],
-			'requirements' => [
-				'segments' => '.+',
-			],
-		]);
-
-		return true;
-	}
-
-	/**
-	 * Register a new route
-	 *
-	 * Route paths can contain wildcard segments, i.e. /blog/owner/{username}
-	 * To make a certain wildcard segment optional, add ? to its name,
-	 * i.e. /blog/owner/{username?}
-	 *
-	 * Wildcard requirements for common named variables such as 'guid' and 'username'
-	 * will be set automatically.
-	 *
-	 * @param string $name   Unique route name
-	 *                       This name can later be used to generate route URLs
-	 * @param array  $params Route parameters
-	 *                       - path : path of the route
-	 *                       - resource : name of the resource view
-	 *                       - defaults : default values of wildcard segments
-	 *                       - requirements : regex patterns for wildcard segment requirements
-	 *                       - methods : HTTP methods
-	 *
-	 * @return Route
-	 * @throws InvalidParameterException
-	 */
-	public function registerRoute($name, array $params = []) {
-
-		$params = $this->hooks->trigger('route:config', $name, $params, $params);
-
-		$path = elgg_extract('path', $params);
-		$resource = elgg_extract('resource', $params);
-		$handler = elgg_extract('handler', $params);
-		$middleware = elgg_extract('middleware', $params, []);
-		$protected = elgg_extract('walled', $params, true);
-
-		if (!$path || (!$resource && !$handler)) {
-			throw new InvalidParameterException(__METHOD__ . ' requires "path" and "resource" parameters to be set');
-		}
-
-		$defaults = elgg_extract('defaults', $params, []);
-		$requirements = elgg_extract('requirements', $params, []);
-		$methods = elgg_extract('methods', $params, []);
-
-		$patterns = [
-			'guid' => '\d+',
-			'group_guid' => '\d+',
-			'container_guid' => '\d+',
-			'owner_guid' => '\d+',
-			'username' => '[\p{L}\p{Nd}._-]+',
-		];
-
-		$path = trim($path, '/');
-		$segments = explode('/', $path);
-		foreach ($segments as &$segment) {
-			// look for segments that are defined as optional with added ?
-			// e.g. /blog/owner/{username?}
-
-			if (!preg_match('/\{(\w*)(\?)?\}/i', $segment, $matches)) {
-				continue;
-			}
-
-			$wildcard = $matches[1];
-			if (!isset($defaults[$wildcard]) && isset($matches[2])) {
-				$defaults[$wildcard] = ''; // make it optional
-			}
-
-			if (!isset($requirements[$wildcard])) {
-				if (array_key_exists($wildcard, $patterns)) {
-					$requirements[$wildcard] = $patterns[$wildcard];
-				} else {
-					$requirements[$wildcard] = '.+?';
-				}
-			}
-
-			$segment = '{' . $wildcard . '}';
-		}
-
-		$path = '/' . implode('/', $segments);
-
-		if ($protected !== false) {
-			$middleware[] = WalledGarden::class;
-		}
-
-		$defaults['_resource'] = $resource;
-		$defaults['_handler'] = $handler;
-		$defaults['_middleware'] = $middleware;
-
-		$route = new Route($path, $defaults, $requirements, [
-			'utf8' => true,
-		], '', [], $methods);
-
-		$this->routes->add($name, $route);
-
-		return $route;
-	}
-
-	/**
-	 * Unregister a route by its name
-	 *
-	 * @param string $name Name of the route
-	 *
-	 * @return void
-	 */
-	public function unregisterRoute($name) {
-		$this->routes->remove($name);
-	}
-
-	/**
-	 * Generate a relative URL for a named route
-	 *
-	 * @param string $name       Route name
-	 * @param array  $parameters Query parameters
-	 *
-	 * @return string
-	 */
-	public function generateUrl($name, array $parameters = []) {
-		try {
-			return $this->generator->generate($name, $parameters, UrlGenerator::ABSOLUTE_URL);
-		} catch (RouteNotFoundException $exception) {
-			elgg_log($exception->getMessage(), 'ERROR');
-
-			return '';
-		}
-	}
-
-	/**
-	 * Populates route parameters from entity properties
-	 *
-	 * @param string          $name       Route name
-	 * @param ElggEntity|null $entity     Entity
-	 * @param array           $parameters Preset parameters
-	 *
-	 * @return array|false
-	 */
-	public function resolveRouteParameters($name, ElggEntity $entity = null, array $parameters = []) {
-		$route = $this->routes->get($name);
-		if (!$route) {
-			return false;
-		}
-
-		$requirements = $route->getRequirements();
-		$defaults = $route->getDefaults();
-		$props = array_merge(array_keys($requirements), array_keys($defaults));
-
-		foreach ($props as $prop) {
-			if (substr($prop, 0, 1) === '_') {
-				continue;
-			}
-
-			if (isset($parameters[$prop])) {
-				continue;
-			}
-
-			if (!$entity) {
-				$parameters[$prop] = '';
-				continue;
-			}
-
-			switch ($prop) {
-				case 'title' :
-				case 'name' :
-					$parameters[$prop] = elgg_get_friendly_title($entity->getDisplayName());
-					break;
-
-				default :
-					$parameters[$prop] = $entity->$prop;
-					break;
-			}
-		}
-
-		return $parameters;
-	}
-
-	/**
-	 * Unregister a page handler for an identifier
-	 *
-	 * @param string $identifier The page type identifier
-	 *
-	 * @return void
-	 * @deprecated 3.0
-	 */
-	public function unregisterPageHandler($identifier) {
-		$this->unregisterRoute($identifier);
-	}
-
-	/**
 	 * Filter a request through the route:rewrite hook
 	 *
 	 * @param Request $request Elgg request
@@ -462,7 +252,7 @@ class Router {
 			'identifier' => $identifier,
 			'segments' => $segments,
 		];
-		$new = _elgg_services()->hooks->trigger('route:rewrite', $identifier, $old, $old);
+		$new = $this->hooks->trigger('route:rewrite', $identifier, $old, $old);
 		if ($new === $old) {
 			return $request;
 		}
